+++ packages/cli/src/playground/src/pages/agents/agent/index.tsx
@@ -1,18 +1,14 @@
-import { PanelLeft } from 'lucide-react';
+import { AgentChat as Chat } from '@mastra/playground-ui';
 import { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router';
 import { v4 as uuid } from '@lukeed/uuid';
-
-import { Chat } from '@/components/Chat';
-import { Button } from '@/components/ui/button';
-import { Skeleton } from '@/components/ui/skeleton';
 
 import { cn } from '@/lib/utils';
 
 import { AgentInformation } from '@/domains/agents/agent-information';
 import { AgentSidebar } from '@/domains/agents/agent-sidebar';
 import { useAgent } from '@/hooks/use-agents';
-import { useMemory, useMessages } from '@/hooks/use-memory';
+import { useMemory, useMessages, useThreads } from '@/hooks/use-memory';
 import type { Message } from '@/types';
 
 function Agent() {
@@ -25,7 +21,12 @@ function Agent() {
     threadId: threadId!,
     memory: !!memory?.result,
   });
-  const [sidebar, setSidebar] = useState(true);
+  const [sidebar, _] = useState(true);
+  const {
+    threads,
+    isLoading: isThreadsLoading,
+    mutate: refreshThreads,
+  } = useThreads({ resourceid: agentId!, agentId: agentId!, isMemoryEnabled: !!memory?.result });
 
   useEffect(() => {
     if (memory?.result && !threadId) {
@@ -36,49 +37,35 @@ function Agent() {
   }, [memory?.result, threadId]);
 
   if (isAgentLoading) {
-    return (
-      <main className="flex-1 relative grid grid-cols-[1fr_400px] divide-x">
-        <div className="p-4">
-          <Skeleton className="h-[600px]" />
-        </div>
-        <div className="flex flex-col">
-          <AgentInformation agentId={agentId!} />
-        </div>
-      </main>
-    );
+    return null;
   }
 
   return (
-    <main
+    <section
       className={cn(
-        'flex-1 relative grid divide-x',
-        sidebar && memory?.result ? 'grid-cols-[256px_1fr_400px] overflow-y-hidden h-full' : 'grid-cols-[1fr_400px]',
+        'relative grid h-full divide-x',
+        sidebar && memory?.result ? 'grid-cols-[256px_1fr_400px]' : 'grid-cols-[1fr_400px]',
       )}
     >
-      {sidebar && memory?.result ? <AgentSidebar agentId={agentId!} threadId={threadId!} /> : null}
-      <div className="relative">
-        {memory?.result ? (
-          <Button
-            variant="primary"
-            size="icon"
-            className="absolute top-4 left-4 z-50"
-            onClick={() => setSidebar(!sidebar)}
-          >
-            <PanelLeft />
-          </Button>
-        ) : null}
+      {sidebar && memory?.result ? (
+        <AgentSidebar agentId={agentId!} threadId={threadId!} threads={threads} isLoading={isThreadsLoading} />
+      ) : null}
+      <div className="relative overflow-y-hidden">
         <Chat
           agentId={agentId!}
           agentName={agent?.name}
           threadId={threadId!}
           initialMessages={isMessagesLoading ? undefined : (messages as Message[])}
           memory={memory?.result}
+          refreshThreadList={() => {
+            refreshThreads();
+          }}
         />
       </div>
       <div className="flex flex-col">
         <AgentInformation agentId={agentId!} />
       </div>
-    </main>
+    </section>
   );
 }