+++ packages/core/src/workflows/types.ts
@@ -89,6 +89,11 @@ export interface BaseCondition<
 }
 
 export type ActionContext<TSchemaIn extends z.ZodType<any>> = IExecutionContext<z.infer<TSchemaIn>, WorkflowContext>;
+export enum WhenConditionReturnValue {
+  CONTINUE = 'continue',
+  CONTINUE_FAILED = 'continue_failed',
+  ABORT = 'abort',
+}
 
 export type StepDef<
   TStepId extends TSteps[number]['id'],
@@ -99,7 +104,12 @@ export type StepDef<
   TStepId,
   {
     snapshotOnTimeout?: boolean;
-    when?: Condition<any, any> | ((args: { context: WorkflowContext; mastra?: MastraPrimitives }) => Promise<boolean>);
+    when?:
+      | Condition<any, any>
+      | ((args: {
+          context: WorkflowContext;
+          mastra?: MastraPrimitives;
+        }) => Promise<boolean | WhenConditionReturnValue>);
     data: TSchemaIn;
     handler: (args: ActionContext<TSchemaIn>) => Promise<z.infer<TSchemaOut>>;
   }
@@ -126,7 +136,10 @@ export interface StepConfig<
   snapshotOnTimeout?: boolean;
   when?:
     | Condition<CondStep, TTriggerSchema>
-    | ((args: { context: WorkflowContext<TTriggerSchema>; mastra?: MastraPrimitives }) => Promise<boolean>);
+    | ((args: {
+        context: WorkflowContext<TTriggerSchema>;
+        mastra?: MastraPrimitives;
+      }) => Promise<boolean | WhenConditionReturnValue>);
   variables?: StepInputType<TStep, 'inputSchema'> extends never
     ? Record<string, VariableReference<VarStep, TTriggerSchema>>
     : {
@@ -141,6 +154,7 @@ type StepSuccess<T> = {
 
 type StepSuspended = {
   status: 'suspended';
+  suspendPayload?: any;
 };
 type StepWaiting = {
   status: 'waiting';
@@ -173,7 +187,7 @@ export type WorkflowEvent =
   | { type: 'RESET_TO_PENDING'; stepId: string }
   | { type: 'CONDITIONS_MET'; stepId: string }
   | { type: 'CONDITION_FAILED'; stepId: string; error: string }
-  | { type: 'SUSPENDED'; stepId: string }
+  | { type: 'SUSPENDED'; stepId: string; suspendPayload?: any }
   | { type: 'WAITING'; stepId: string }
   | { type: `xstate.error.actor.${string}`; error: Error }
   | { type: `xstate.done.actor.${string}`; output: ResolverFunctionOutput };
@@ -195,6 +209,7 @@ export type SubscriberFunctionOutput = {
 
 export type DependencyCheckOutput =
   | { type: 'CONDITIONS_MET' }
+  | { type: 'CONDITIONS_SKIPPED' }
   | { type: 'CONDITION_FAILED'; error: string }
   | { type: 'SUSPENDED' }
   | { type: 'WAITING' };